Refactors first implementation of graphql endpoint
